+++ addon/components/flexberry-objectlistview.js
@@ -111,6 +111,14 @@ export default FlexberryBaseComponent.extend({
   }),
 
   /**
+    Set the attribute name to hierarchy build.
+
+    @property hierarchyAttribute
+    @type String
+  */
+  hierarchyAttribute: undefined,
+
+  /**
     The name of a property in the model that determines whether any record is the parent of other records.
     If a property value with this name is defined (not `undefined`), the button to display child records will be shown immediately, and the records will be loaded only when the button is clicked.