+++ blueprints/default-user-setting/files/__root__/default-user-settings/__name__.js
@@ -1,0 +1,3 @@
+export default class UserSetting {}
+
+UserSetting.DEFAULT = {};