+++ CHANGELOG.md
@@ -6,6 +6,25 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 * Replace test models localization with separated classes.
 
+## [2.6.0-beta.4] - 2021-01-15
+### Added
+* The `indeterminate` mode for `flexberry-checkbox` by setting `isNullable`.
+
+### Fixed
+* The `default-user-setting` blueprint:
+    * Fixed init component path.
+
+## [2.6.0-beta.3] - 2021-01-13
+### Fixed
+* The `flexberry-objectlitview` component:
+    * Add ability to on and off hierarchical mode when component is in hierarchical mode initially.
+
+## [2.6.0-beta.2] - 2020-12-20
+### Added
+* The `isDesktop` and `isMobile` methods for `device` service.
+* The `hierarchyAttribute` property for `flexberry-objectlistview` component.
+* Ability to define default user settings for the type of models to be used in the lists.
+
 ## [2.6.0-beta.1] - 2020-11-15
 ### Added
 * The `hasShadow` and `hasBorder` properties for `flexberry-toggler` component.